ai/requirements.txt
langchain==0.2.15
langchain-openai==0.1.23
langchain-core==0.2.35
langchain-community==0.2.13
fastapi==0.112.2
uvicorn==0.30.6
python-dotenv==1.0.1
langfuse==2.44.1
pydantic==2.8.2
openai==1.42.0
typing-extensions==4.12.2
aiohttp==3.10.5
httpx==0.27.2 